refactor(nextjs): migrate from src/realpages pattern on `.real` suffix
