+++ nextjs/home/pages/testRemoteHook.js
@@ -1,7 +1,8 @@
 import dynamic from 'next/dynamic';
-const page = import('./index.real');
 
-const Page = dynamic(() => import('./index.real'));
+const page = import('./testRemoteHook.real');
+const Page = dynamic(() => import('./testRemoteHook.real'), { ssr: false });
+
 Page.getInitialProps = async ctx => {
   const getInitialProps = (await page).default?.getInitialProps;
   if (getInitialProps) {
@@ -9,4 +10,5 @@ Page.getInitialProps = async ctx => {
   }
   return {};
 };
+
 export default Page;
