+++ nextjs/shop/components/useCustomRemoteHook.js
@@ -1,10 +1,12 @@
 import { useState, useEffect } from 'react';
 
-function customHook(friendID) {
+function useCustomRemoteHook(friendID) {
   const [isOnline, setIsOnline] = useState(null);
 
   useEffect(() => {
     console.log('some custom hook');
   }, []);
+
+  return 'Custom hook from localhost:3002 works!';
 }
-export default customHook;
+export default useCustomRemoteHook;
