+++ nextjs/shop/pages/index.js
@@ -1,12 +1,3 @@
-import dynamic from 'next/dynamic';
-const page = import('../realPages/index');
+import { createFederatedCatchAll } from 'nextjs-shared';
 
-const Page = dynamic(() => import('../realPages/index'));
-Page.getInitialProps = async ctx => {
-  const getInitialProps = (await page).default?.getInitialProps;
-  if (getInitialProps) {
-    return getInitialProps(ctx);
-  }
-  return {};
-};
-export default Page;
+export default createFederatedCatchAll(['checkout', 'home']);
