--- nextjs/shop/realPages/[...federatedPage].js
@@ -1,3 +1,0 @@
-import { createFederatedCatchAll } from 'nextjs-shared';
-
-export default createFederatedCatchAll();
