--- nextjs/shop/realPages/p/[...slug].js
@@ -1,6 +1,0 @@
-export default function PDP() {
-  return <h1>PDP!!!</h1>;
-}
-PDP.getInitialProps = async () => {
-  return {};
-};
